+++ CarLoop/app/src/main/java/com/buf/carloop/MessageClass.java
@@ -92,7 +92,7 @@ public class MessageClass {
 
         task.execute(sqlComm);
         try {
-            int value = task.get(100000, TimeUnit.MILLISECONDS);
+            int value = (int) task.get(100000, TimeUnit.MILLISECONDS);
             return value;
         } catch (Exception e) {
             e.printStackTrace();
