+++ CarLoop/app/src/main/java/com/buf/carloop/Vehicle.java
@@ -149,7 +149,7 @@ public class Vehicle {
         AsyncSQLLongHaul task = new AsyncSQLLongHaul();
         task.execute(sqlComm);
         try {
-            return task.get(10000, TimeUnit.MILLISECONDS);
+            return (int) task.get(10000, TimeUnit.MILLISECONDS);
         } catch (Exception e) {
             e.printStackTrace();
             return -1;
@@ -170,7 +170,7 @@ public class Vehicle {
         AsyncSQLLongHaul task = new AsyncSQLLongHaul();
         task.execute(sqlComm);
         try {
-            return task.get(10000, TimeUnit.MILLISECONDS);
+            return (int) task.get(10000, TimeUnit.MILLISECONDS);
         } catch (Exception e) {
             e.printStackTrace();
             return -1;
